Move filename of KeyInfo
